requirements.txt
robotframework==3.1.1
confluent-kafka==1.0.0
uuid==1.30
